+++ svc/pkg/job-log/db/log/migrations/20200101000000_init.up.sql
@@ -1,0 +1,14 @@
+CREATE TABLE IF NOT EXISTS run_logs
+(
+	run_id UUID,
+	task LowCardinality(String),
+	stream_type UInt8,  -- backend::job::log::StreamType
+	ts DateTime64(9),
+	message String
+)
+ENGINE = ReplicatedMergeTree()
+PARTITION BY toStartOfHour(ts)
+ORDER BY (run_id, toUnixTimestamp(ts), stream_type)
+TTL toDate(ts + toIntervalDay(3))
+SETTINGS index_granularity = 8192, ttl_only_drop_parts = 1;
+