+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   **Infra** Turnstile `CAPTCHA_CAPTCHA_REQUIRED` responses now include a site key
 -   **Infra** Turnstile is no longer configurable by domain (instead configured by Turnstile itself)
 -   **Infra** Fixed /list not returning lobbies unless the `include_state` query parameter was true
+-   **Infra** Job log aggregating to use Vector under the hood to insert directly in to ClickHouse
+
+### Fixed
+
+-   **Infra** Job logs occasionally returning duplicate log lines
 
 ## [23.1.0] - 2023-10-30